Rename test/package.js to test/compiler.js
