+++ test/compiler.js
