+++ test/index.js
@@ -2,7 +2,7 @@ const semver = require('semver');
 
 require('./linker.js');
 require('./translate.js');
-require('./package.js');
+require('./compiler.js');
 require('./smtchecker.js');
 require('./abi.js');
 require('./determinism.js');
